A collection of minimal Examples from various designs with one or two factors.

The datasets present minimal examples that are analyzed with an Analysis of Frequency Data method (described in lc23;textualANOPA. The five datasets are

'minimalBSExample': an example with a single factor (state of residency)

'twoWayExample': an example with two factors, Class and Difficulty

'minimalWSExample': an example with a within-subject design (three measurements)

'twoWayWithinExample': an example with two within-subject factors

'minimalMxExample': a mixed design having one within and one between-subject factors

'minimalMxExampleCompiled': a mixed design having one within and one between-subject factors but available in a compiled format (more compact).

Usage

minimalBSExample

twoWayExample

minimalWSExample

twoWayWithinExample

minimalMxExample

minimalMxExampleCompiled

Format

Objects of class data.frame:

An object of class data.frame with 6 rows and 4 columns.

An object of class data.frame with 19 rows and 3 columns.

An object of class data.frame with 30 rows and 6 columns.

An object of class data.frame with 27 rows and 5 columns.

An object of class data.frame with 4 rows and 5 columns.

References

Examples

library(ANOPA)

# the twoWayExample data with proportions per Classes and Difficulty levels 
twoWayExample
#>   Class Difficulty success total
#> 1 First       Easy      11    12
#> 2 First   Moderate       9    12
#> 3 First  Difficult       6    12
#> 4  Last       Easy      10    12
#> 5  Last   Moderate       8    12
#> 6  Last  Difficult       3    12

# perform an anopa on this dataset
w <- anopa( {success;total} ~ Difficulty * Class, twoWayExample) 

# We analyse the proportions by Difficulty for each Class
e <- emProportions(w, ~ Difficulty | Class)