contrastProportion: analysis of contrasts between proportions using Anscombe transform.

The function contrastProportions() performs contrasts analyses on proportion data after an omnibus analysis has been obtained with anopa() according to the ANOPA framework. See Laurencelle and Cousineau (2023) for more.

contrastProportions(w = NULL, contrasts = NULL)

Arguments

w

An ANOPA object obtained from anopa() or emProportions();

contrasts

A list that gives the weights for the contrasts to analyze. The contrasts within the list can be given names to distinguish them. The contrast weights must sum to zero and their cross-products must equal 0 as well.

Value

A table of significance of the different contrasts.

Details

contrastProportions() computes the _F_s for the contrasts, testing the hypothesis that it equals zero. The contrasts are each 1 degree of freedom, and the sum of the contrasts' degrees of freedom totalize the effect-being-decomposed's degree of freedom.

References

Laurencelle L, Cousineau D (2023). “Analysis of proportions using arcsine transform with any experimental design.” Frontiers in Psychology, 13, 1045436. doi:10.3389/fpsyg.2022.1045436 .

Examples

# Basic example using a one between-subject factor design with the data in compiled format. 
# Ficticious data present success or failure of observation classified according
# to the state of residency (three levels); 175 participants have been observed in total.

# The cells are unequal:
minimalBSExample
#>      state  s  n
#> 1  Florida 31 57
#> 2 Kentucky 25 73
#> 3  Montana  9 45

# First, perform the omnibus analysis :
w <- anopa( {s;n} ~ state, minimalBSExample) 
summary(w)
#>             MS  df        F        p correction   Fcorr pvalcorr
#> state 0.032384   2 7.335621 0.000652   1.011881 7.24949 0.000711
#> Error 0.004415 Inf                                              

# Compare the first two states jointly to the third, and
# compare the first to the second state:
cw <- contrastProportions( w, list(
         contrast1 = c(1,  1, -2)/2,
         contrast2 = c(1, -1,  0) )
      )
#> Not yet programmed...
#summary(cw)

# Example using the Arrington et al. (2002) data, a 3 x 4 x 2 design involving 
# Location (3 levels), Trophism (4 levels) and Diel (2 levels).
ArringtonEtAl2002
#>                 Location    Trophism      Diel    s    n
#> 1                 Africa Detritivore   Diurnal   16  217
#> 2                 Africa Invertivore   Diurnal   76  498
#> 3                 Africa Invertivore Nocturnal   55  430
#> 4                 Africa    Omnivore   Diurnal    2   87
#> 5                 Africa   Piscivore   Diurnal  673  989
#> 6                 Africa   Piscivore Nocturnal  221  525
#> 7  Central/South America Detritivore   Diurnal   68 1589
#> 8  Central/South America Detritivore Nocturnal    9  318
#> 9  Central/South America Invertivore   Diurnal  706 7452
#> 10 Central/South America Invertivore Nocturnal  486 2101
#> 11 Central/South America    Omnivore   Diurnal  293 6496
#> 12 Central/South America    Omnivore Nocturnal   82  203
#> 13 Central/South America   Piscivore   Diurnal 1275 5226
#> 14 Central/South America   Piscivore Nocturnal  109  824
#> 15         North America Detritivore   Diurnal  142 1741
#> 16         North America Invertivore   Diurnal  525 3368
#> 17         North America Invertivore Nocturnal  231 1539
#> 18         North America    Omnivore   Diurnal  210 1843
#> 19         North America    Omnivore Nocturnal    7   38
#> 20         North America   Piscivore   Diurnal  536 1289
#> 21         North America   Piscivore Nocturnal   19  102

# performs the omnibus analysis first (mandatory):
w <- anopa( {s;n} ~ Location * Trophism * Diel, ArringtonEtAl2002) 
#> ANOPA::fyi(1): Combination of cells missing. Adding: 
#>       Location    Trophism      Diel s n
#>         Africa Detritivore Nocturnal 0 0
#>         Africa    Omnivore Nocturnal 0 0
#>  North America Detritivore Nocturnal 0 0
#> Warning: ANOPA::warning(1): Some cells have zero over zero data. Imputing...
corrected(w)
#>                              MS  df        F correction    Fcorr pvalcorr
#> Location               0.027449   2 0.961802   1.000112 0.961694 0.382245
#> Trophism               0.095656   3 3.351781   1.000115 3.351396 0.018111
#> Diel                   0.029715   1 1.041227   1.000049 1.041176 0.307548
#> Location:Trophism      0.029485   6 1.033146   1.013842 1.019041 0.410515
#> Location:Diel          0.005277   2 0.184900   1.010164 0.183040 0.832735
#> Trophism:Diel          0.073769   3 2.584868   1.012197 2.553721 0.053559
#> Location:Trophism:Diel 0.011297   6 0.395837   1.055660 0.374967 0.895351
#> Error(between)         0.028539 Inf                                      

# execute the simple effect of Trophism for every levels of Diel and Location
e <- emProportions(w, ~ Trophism | Diel * Location)
#> Not yet programmed...
#summary(e)

# For each of these sub-cells, contrast the four tropisms, first
# by comparing the first two levels to the third (contrast1), second
# by comparing the first to the second level (contrast2), and finally by
# by comparing the first three to the last (contrast3) :
#f <- contrastProportions( e, list(
#              contrast1 = c(1,  1, -2, 0)/2,
#              contrast2 = c(1, -1,  0, 0),
#              contrast3 = c(1,  1,  1, -3)/3
#           )
#      )
#summary(f)